Add X-axis labels to EntriesChart component
Enhanced the EntriesChart.razor component by adding X-axis labels to the chart. Introduced a new private field `_xAxisLabels` to hold the X-axis labels, which are populated with dates from the `Entries` collection formatted as short date strings. The `XAxisLabels` property of the `MudChart` component is set to use this new field, making the chart more informative and readable.